Add flow filtering on graph
Mask the edges of the graph that should not be affected by the solution.
Edges are identified by edge IDs for now. BUT...

The mask is set after the flows are split. Thus the edge IDs that we
take from the DB are duplicated on the flows that are split. That's
why the edges should be identified by (source, target, material) instead
of edge IDs